requirements.txt
confluent-kafka[avro]>=2.4.0
python-dotenv>=1.0.1
Flask>=3.0.0
langchain>=0.3.27
langchain-openai>=0.3.32
langchain-mcp-adapters>=0.1.9
langgraph>=0.6.6
mcp>=1.13.1
neo4j>=5.23.1
pypdf>=4.3.1
